# backend/doc/openapi.yaml
openapi: 3.0.1
info:
  title: Stupro 2019 Querimonia API
  description: Querimonia ist eine Plattform zum automatischen Beschwerdemanagement. Über die hier beschriebene API kann mit dem Backend kommuniziert werden. Dies beinhaltet Upload von neuen Beschwerden, Abrufen der gespeicherten Beschwerden, Anfragen von Antworttexten, etc.
  version: 0.5.2
servers:
  - url: https://beschwerdemanagement.iao.fraunhofer.de/dev/
  - url: https://beschwerdemanagement.iao.fraunhofer.de/
  - url: https://querimonia.iao.fraunhofer.de/dev
  - url: https://querimonia.iao.fraunhofer.de/
  - url: http://localhost:8080/
...
paths:
  /api/complaints:
    get:
      tags:
        - complaints
      summary: Gibt Beschwerden zurück, die auf dem Server gespeichert sind.
      description: Gibt Beschwerden zurück, die in der Datenbank gespeichert wurden. Es können über die Parameter Filter angewandt werden, die Sortierung angegeben werden und die Anzahl und das Offset der Beschwerden bestimmt werden.
      parameters:
        - $ref: '#/components/parameters/CountParameter'
        - $ref: '#/components/parameters/PageParameter'
        - $ref: '#/components/parameters/ComplaintsSortParameter'
        - $ref: '#/components/parameters/ComplaintStateParameter'
        - $ref: '#/components/parameters/DateMinParameter'
        - $ref: '#/components/parameters/DateMaxParameter'
        - $ref: '#/components/parameters/SentimentParameter'
        - $ref: '#/components/parameters/SubjectParameter'
        - $ref: '#/components/parameters/KeywordsParameter'

      operationId: getComplaints
      responses:
        200:
          description: Erfolgreiche Anfrage.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Complaint'
        400:
          description: Üngültige Angabe von Parametern.
        5XX:
          description: Ein unerwarteter Server Fehler ist aufgetreten.
  /api/complaints/import:
    post:
      tags:
        - complaints
      summary: Läd eine neue Beschwerde in die Datenbank.
      description: Über diese Schnittstelle können Dateien hochgeladen werden. Akzeptiert werden txt, docx, doc und pdf Dateien mit einer maximalen Zeichenanzahl von 5000 Zeichen. Alternativ können über application/json direkt Texte eingefügt werden. Die Beschwerde wird beim Upload analysiert und eine Antwort wird generiert. Über den Query-Parameter configId kann eine bestimmte Konfiguration angegeben werden, mit der die neue Beschwerde analysiert werden soll.
      parameters:
        - $ref: '#/components/parameters/ConfigIdQueryParameter'
      requestBody:
        description: Die Datei, die hochgeladen werden soll.
        content:
          multipart/form-data:
            schema:
              type: string
              format: binary
              example: file content
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: string
                  example: "Das ist eine Beispielbeschwerde"
      operationId: importComplaint
      responses:
        201:
          description: Datei wurde erfolgreich hochgeladen, der Text extrahiert, ausgewertet und als Beschwerde im System gespeichert.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Complaint'
        400:
          description: Text konnte nicht aus Datei extrahiert werden oder Text ist zu lang.
        500:
          description: Fehlerhafte Konfiguration der Extraktoren, Klassifikatoren oder den Antwortbausteinen oder externe Extraktoren konnten nicht aufgerufen werden.
  /api/complaints/{complaintId}:
    parameters:
      - $ref: '#/components/parameters/ComplaintIdParameter'

    get:
      tags:
        - complaints
      summary: Gibt eine Beschwerde mit der gegebenen Id zurück.
      description: Gibt eine Beschwerde mit der gegebenen Id zurück.
      operationId: getComplaint
      responses:
        200:
          description: Erfolgreiche Anfrage.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Complaint'
        404:
          description: Beschwerde mit gegebener Id existiert nicht.

    patch:
      tags:
        - complaints
      summary: Aktualisiert eine Beschwerde.
      description: Mit dieser Anfrage können Sentiment, Kategorie und Status einer Beschwerde aktualisiert werden. Werden Kategorie und Sentiment gändert, so wird das setByUser-Attribut auf true gesetzt.
      requestBody:
        description: Ein Objekt mit Attributen, die aktualisiert werden sollen. Es müssen nur die Attribute angegeben werden, bei denen eine Änderung an der Beschwerde vorgenommen werden soll.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComplaintUpdateRequest'
            example:
              "state": "CLOSED"
              "subject": "Fahrt nicht erfolgt"
      operationId: updateComplaint
      responses:
        200:
          description: Beschwerde wurde erfolgreich aktualisiert
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Complaint'

    delete:
      tags:
        - complaints
      summary: Löscht eine Beschwerde.
      description: Löscht eine Beschwerde mit gegebener Id, falls diese vorhanden ist.
      operationId: deleteComplaint
      responses:
        204:
          description: Löschen erfolgreich.
        404:
          description: Beschwerde mit gegebener Id existiert nicht
  /api/complaints/{complaintId}/refresh:
    parameters:
      - $ref: '#/components/parameters/ComplaintIdParameter'

    patch:
      tags:
        - complaints
      summary: Ruft alle Analyse-Funktionen erneut auf.
      description: Alle Tools werden neu ausgeführt, vorhandene Informationen werden überschrieben. Über die Parameter kann definiert werden, ob vom Nutzer gesetzte Information dabei nicht überschreiben werden soll und welche Konfiguration verwendet werden soll.
      parameters:
        - name: keepUserInformation
          in: query
          description: Falls dieser Wert auf true gesetzt wird, werden vom Nutzer gesetzte Informationen nicht überschrieben.
          required: false
          schema:
            type: boolean
            default: false
        - $ref: '#/components/parameters/ConfigIdQueryParameter'
      operationId: refreshComplaint
      responses:
        200:
          description: Beschwerde wurde erfolgreich aktualisiert
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Complaint'
  /api/complaints/count:
    get:
      tags:
        - complaints
      summary: Gibt die Anzahl der Beschwerden zurück.
      description: Gibt die Anzahl von Beschwerden in der Datenbank zurück, wobei Filter angewendet werden können.
      parameters:
        - $ref: '#/components/parameters/DateMinParameter'
        - $ref: '#/components/parameters/DateMaxParameter'
        - $ref: '#/components/parameters/SentimentParameter'
        - $ref: '#/components/parameters/SubjectParameter'
        - $ref: '#/components/parameters/KeywordsParameter'

      operationId: countComplaints
      responses:
        200:
          description: Erfolgreiche Anfrage.
          content:
            text/plain:
              schema:
                type: integer
                format: int32
                example: 140
        400:
          description: Üngültige Angabe von Parametern.
  /api/complaints/{complaintId}/entities:
    parameters:
      - $ref: '#/components/parameters/ComplaintIdParameter'

    post:
      tags:
        - complaints
      summary: Fügt eine neue Named Entity hinzu.
      description: Eine neue Named Entity wird der Liste der Entitäten der Beschwerde hinzugefügt.
      parameters:
        - $ref: '#/components/parameters/LabelParameter'
        - $ref: '#/components/parameters/StartParameter'
        - $ref: '#/components/parameters/EndParameter'
        - $ref: '#/components/parameters/ExtractorParameter'

      operationId: addEntity
      responses:
        200:
          description: Die Entity wurde der Beschwerde hinzugefügt. Eine aktualisierte Entity-Liste wird zurückgegeben.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NamedEntity'
        400:
          description: Die gegebene Entität ist schon vorhanden (mit gleichem Extraktor an der exakt gleichen Position), der gebene Extraktor ist unbekannt, die Beschwerde ist bereits abgeschlossen oder die gegebenen Indices liegen außerhalb der Textgrenzen.

    delete:
      tags:
        - complaints
      summary: Löscht eine Named Entity.
      description: Eine neue Named Entity wird aus der Liste der Entitäten entfernt.
      parameters:
        - $ref: '#/components/parameters/LabelParameter'
        - $ref: '#/components/parameters/StartParameter'
        - $ref: '#/components/parameters/EndParameter'
        - $ref: '#/components/parameters/ExtractorParameter'

      operationId: deleteEntity
      responses:
        200:
          description: Die Entity wurde der Beschwerde hinzugefügt. Eine aktualisierte Entity-Liste wird zurückgegeben.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NamedEntity'
        404:
          description: Die gegebene Entität ist nicht vorhanden.
...
  /api/templates:
    get:
      tags:
        - templates
      summary: Gibt die gespeicherten Antwortbausteine zurück.
      description: Hier können alle in der Datenbank gespeicherten Antwortbausteine abgerufen werden. Dabei ist Filtern/Pagination möglich.
      parameters:
        - $ref: '#/components/parameters/CountParameter'
        - $ref: '#/components/parameters/PageParameter'
        - name: sort_by
          in: query
          description: Hier können alle Kategorien angegeben werden, nach denen die Anfrage sortiert werden soll. Die Reihenfolge im Array bestimmt die Priorität beim Sortieren.
          required: false
          schema:
            type: array
            items:
              type: string
              enum: [id_desc, id_asc, priority_desc, priority_asc, name_desc, name_asc]
              default: name_asc
              example: name_asc
        - $ref: '#/components/parameters/KeywordsParameter'

      operationId: getResponseComponents
      responses:
        200:
          description: Die Antwortbausteine werden gefiltert und sortiert zurückgegeben.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResponseComponent'
        400:
          description: Fehlerhafte Parameter.

    post:
      tags:
        - templates
      summary: Fügt dem System einen neuen Antwortbaustein hinzu.
      description: Fügt einen neuen Antwortbaustein hinzu. Dieser kann Platzhalter enthalten, welche mit Entitäten aus der Beschwerden ersetzt werden.
      requestBody:
        description: Der neue Textbaustein, der dem System hinzufügt werden soll. Im Request-Body müssen ID und requiredEntities nicht angegeben werden, diese werden bei Angabe nicht beachtet.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResponseComponent'

      operationId: addResponseComponent
      responses:
        201:
          description: Das Template ist richtig formatiert, die Regelbeschreibung ist gültig und wurde in der Datenbank gespeichert.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseComponent'
        400:
          description: Der Textbaustein enthält ungültige Formatierungen, die Regelbeschreibung ist ungültig oder überschreitet das Zeichenlimit von 5000 Zeichen.
...
components:
  parameters:
...
    ComplaintStateParameter:
      name: state
      in: query
      required: false
      description: Gibt die Status an, in dem sich eine Beschwerde befindet. Ohne Angabe werden geschlossene Beschwerden nicht zurückgegeben.
      schema:
        type: array
        items:
          $ref: '#/components/schemas/ComplaintState'
        default:
          - NEW
          - IN_PROGRESS
...
    CountParameter:
      name: count
      in: query
      description: Anzahl der Elemente, die zurück gegeben werden sollen. Falls nicht angegeben, werden alle Elemente zurückgegeben.
      required: false
      schema:
        type: integer
        format: int32
        minimum: 1
        example: 10
    PageParameter:
      name: page
      in: query
      description: Falls page 0 ist, werden die ersten *count* Elemente zurück gegeben, falls page 1 ist die zweiten *count* Elemente, gibt also die Seitenzahl an. Falls nicht angegeben, wird 0 angenommen. Wird ignoriert, falls *count* nicht gegeben ist.
      required: false
      schema:
        type: integer
        format: int32
        default: 0
        minimum: 0
        example: 0
    ComplaintsSortParameter:
      name: sort_by
      in: query
      required: false
      description: Hier können alle Kategorien angegeben werden, nach denen die Anfrage sortiert werden soll. Die Reihenfolge im Array bestimmt die Priorität beim Sortieren.
      schema:
        description: Ein Array von Kategorien, nach denen eine Anfrage sortiert werden soll. Zusätzlich wird angegeben, ob auf oder absteigend sortiert wird.
        type: array
        items:
          type: string
          description: Gibt an, nach was die Beschwerden sortiert werden sollen.
          enum:
            - upload_date_asc
            - upload_date_desc
            - subject_asc
            - subject_desc
            - sentiment_asc
            - sentiment_desc
        default: ["upload_date_desc"]
    DateMinParameter:
      name: date_min
      in: query
      required: false
      description: Bei Angabe werden nur Beschwerden ausgewertet, die nach diesem Datum hochgeladen wurden. Datum muss [RFC3339 5.6](https://tools.ietf.org/html/rfc3339#section-5.6) enstprechen, also dem Format yyyy-mm-dd.
      schema:
        type: string
        format: date
        example: 2019-03-29
    DateMaxParameter:
      name: date_max
      in: query
      required: false
      description: Bei Angabe werden nur Beschwerden ausgewertet, die vor diesem Datum hochgeladen wurden. Das Datum entpricht dem gleichen Format wie bei *dateMin*
      schema:
        type: string
        format: date
        example: 2019-07-23
    SentimentParameter:
      name: sentiment
      in: query
      required: false
      description: Bei Angabe werden nur Beschwerden ausgewertet, die einem der gegebenen Sentimenten entsprechen.
      schema:
        type: array
        items:
          type: string
        example:
          - "Wut"
          - "Unzufriedenheit"
    SubjectParameter:
      name: subject
      in: query
      required: false
      description: Bei Angabe werden nur Beschwerden ausgewertet, die einem der gegebenen Kategorien von Beschwerden entsprechen.
      schema:
        type: array
        items:
          type: string
        example:
          - "Fahrt nicht erfolgt"
          - "Sonstiges"
    KeywordsParameter:
      name: keywords
      in: query
      required: false
      description: Bei Angabe werden nur Elemente ausgewertet, die die angebenen Schlüsselwörter enthalten.
      schema:
        type: array
        items:
          type: string
        example:
          - "Busfahrer"
          - "unfreundlich"
    ConfigIdQueryParameter:
      name: configId
      in: query
      description: Falls dieser Wert gegeben wird, wird die Beschwerde mit der gegebenen Konfiguration analysiert. Ansonsten wird die aktuell ausgewählt Konfiguration ausgewählt.
      required: false
      schema:
        type: integer
        format: int32
        minimum: 1
        example: 1
...
  schemas:
    Complaint:
      required:
        - state
        - configuration
        - text
        - preview
        - receiveDate
        - receiveTime
        - subject
        - sentiment
        - entities
      type: object
      description: Diese Datenstuktur modelliert eine Beschwerde (oder eine eingegangen Meldung von anderer Art). Eine Beschwerde besitzt eine eindeutige ID, einen Zustand (in Bearbeitung, geschlossen), den Meldetext, Eingangsdatum, eine Kategorie und Sentiment so wie eine Liste an erkannten Entitäten. Weiterhin ist zu jeder Beschwerde gespeichert mit welcher Konfigutartion sie analysiere wurde,
      properties:
        complaintId:
          type: integer
          format: int32
          minimum: 1
          example: 1
          description: Die eindeutige ID der Beschwerde.
        state:
          $ref: '#/components/schemas/ComplaintState'
        configuration:
          $ref: '#/components/schemas/AnalyzerConfiguration'
        text:
          type: string
          description: Der volle Beschwerdetext.
          example: "Guten Tag,\nMein Bus kam heute viel zu spät!\nSie müssen mir den Schaden ersetzen!!"
        preview:
          type: string
          description: Die ersten zwei Zeilen der Beschwerde.
          example: "Guten Tag,\nMein Bus kam heute viel zu spät!"
        receiveDate:
          type: string
          format: date
          description: Das Eingangsdatum der Beschwerde im Format yyyy-mm-dd.
          example: "2019-05-02"
        receiveTime:
          type: string
          format: date-time
          description: Die Eingangszeit der Beschwerde im System im Format hh:mm:ss
          example: "13:45:00"
        subject:
          $ref: '#/components/schemas/ComplaintProperty'
        sentiment:
          $ref: '#/components/schemas/ComplaintProperty'
        entities:
          type: array
          items:
            $ref: '#/components/schemas/NamedEntity'
          description: Enthält alle Entitäten, die in der Beschwerde gefunden wurden.
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic